Add support for nanocmd workflows
